update to use 5.0.7 and no haproxy

+++ fabfile.py
@@ -3,12 +3,13 @@
 Fabric script to manage demo
 """
 
-from fabric.api import env, cd, sudo, run
+from fabric.api import env, cd, sudo
 from fabric.decorators import task
 
-# Fabric uses the ~/.ssh/config 
+# Fabric uses the ~/.ssh/config
 # Needed here because of wrong dns
 env.use_ssh_config = True
+
 
 @task
 def demo_host(branch='master'):
@@ -56,18 +57,17 @@ def setup():
     with cd(env.homedir):
 
         # clone repository from github
-        sudo('git clone https://github.com/collective/demo.plone.de.git', user=env.deploy_user)
-
+        sudo('git clone https://github.com/collective/demo.plone.de.git', user=env.deploy_user)  # noqa: E501
 
     with cd(env.directory):
 
         # requirements
-        # sudo('python python-dev build-essential zlib1g-dev libssl-dev libxml2-dev libxslt1-dev wv poppler-utils libtiff5-dev libjpeg62-dev zlib1g-dev libfreetype6-dev liblcms1-dev libwebp-dev')
+        # sudo('python python-dev build-essential zlib1g-dev libssl-dev libxml2-dev libxslt1-dev wv poppler-utils libtiff5-dev libjpeg62-dev zlib1g-dev libfreetype6-dev liblcms1-dev libwebp-dev')   # noqa: E501
 
         # prepare buildout
         sudo('ln -s local_production.cfg local.cfg', user=env.deploy_user)
-        sudo('echo -e "[buildout]\nlogin = admin\npassword = admin" > secret.cfg', user=env.deploy_user)
-        
+        sudo('echo -e "[buildout]\nlogin = admin\npassword = admin" > secret.cfg', user=env.deploy_user)  # noqa: E501
+
         # bootstrap and run bildout once
         sudo('python bootstrap.py', user=env.deploy_user)
         sudo('./bin/buildout', user=env.deploy_user)
@@ -82,25 +82,32 @@ def update():
     Update the instance and reinstall the demo
     """
     with cd(env.directory):
-        
-        # stop running plone
-        stop()
-        
-        # update plone
-        sudo('git pull', user=env.deploy_user)
-        sudo('git checkout {}'.format(env.branch), user=env.deploy_user)
-        
-        # bootstrap
-        sudo('python bootstrap.py', user=env.deploy_user)
-        
-        # delete database and sone files
-        sudo('rm -rf .installed.cfg', user=env.deploy_user)
-        sudo('rm -rf ./var/blobstorage', user=env.deploy_user)
-        sudo('rm -rf ./var/filestorage', user=env.deploy_user)        
 
-        # buildout
-        sudo('./bin/buildout', user=env.deploy_user)
-        
-        # start zope
-        start()
-        sudo('/usr/bin/wget -S -qO- demo.plone.de > /tmp/demo.plone.de.html', user=env.deploy_user)
+        # update plone
+        result = sudo('git pull', user=env.deploy_user)
+        quick_update = 'Already up-to-date.' in result
+
+        if quick_update:
+            # Plonesite Recipe replaces site on the fly
+            print 'UPDATE: No full Buildout required: {0:s}'.format(result)
+            # buildout
+            sudo('./bin/buildout install plonesite', user=env.deploy_user)
+
+        else:
+            stop()
+            sudo('git checkout {}'.format(env.branch), user=env.deploy_user)
+
+            # bootstrap
+            sudo('python bootstrap.py', user=env.deploy_user)
+
+            sudo('rm -rf ./var/blobstorage', user=env.deploy_user)
+            sudo('rm -rf ./var/filestorage', user=env.deploy_user)
+
+            # buildout
+            sudo('./bin/buildout', user=env.deploy_user)
+            sudo('./bin/zeoclient_debug adduser admin admin', user=env.deploy_user)  # noqa: E501
+
+            # start zope
+            start()
+
+        sudo('/usr/bin/wget -S -qO- demo.plone.de > /tmp/demo.plone.de.html', user=env.deploy_user)  # noqa: E501
